Various fixed for Windows build
Qmake uses a different folder structure on Windows. Inserted some
ifdefs into build scripts.

Also some sompilation errors in boost, which haven't shown up before
added gcc pragmas to supress them. This should ideally be handled in
a different way, but I don't have time for this right now.

+++ nmlib/model/module.cpp
@@ -84,7 +84,7 @@ const InputLink *Module::getInput(unsigned int i) const
 
 unsigned int Module::getInputSize() const
 {
-    return static_cast<unsigned int>(m_inputs.size());
+    return m_inputs.size();
 }
 
 std::vector<InputLink *> Module::getInputs()
@@ -136,7 +136,7 @@ OutputLink *Module::getOutput(unsigned int i)
 
 unsigned int Module::getOutputSize() const
 {
-    return static_cast<unsigned int>(m_outputs.size());
+    return m_outputs.size();
 }
 
 std::vector<OutputLink *> Module::getOutputs()
